refactor clientjs file
