+++ desktop/apps/auction2/client/filter_query.js
