+++ desktop/apps/auction2/components/bid_status/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'
 
 export default function BidStatus({ saleArtwork }) {
   let bidLabel
