+++ desktop/apps/auction2/components/filter_sort/index.js
@@ -1,4 +1,4 @@
-import { updateSort } from '../../actions'
+import { updateSort } from '../../client/actions'
 import React from 'react';
 import { connect } from 'react-redux'
 import _ from 'underscore'
