+++ desktop/apps/auction2/components/medium_filter/index.js
@@ -1,4 +1,4 @@
-import { updateMediumParams } from '../../actions'
+import { updateMediumParams } from '../../client/actions'
 import BasicCheckbox from '../basic_checkbox'
 import React from 'react';
 import { connect } from 'react-redux'
