+++ desktop/apps/auction2/components/works_by_followed_artists/index.js
@@ -1,9 +1,9 @@
-import { nextPageOfFollowedArtistArtworks, previousPageOfFollowedArtistArtworks } from '../../actions'
+import { nextPageOfFollowedArtistArtworks, previousPageOfFollowedArtistArtworks } from '../../client/actions'
 import classNames from 'classnames'
 import AuctionGridArtwork from '../auction_grid_artwork'
 import ChevronLeft from '../../../../components/main_layout/public/icons/chevron-left.svg'
 import ChevronRight from '../../../../components/main_layout/public/icons/chevron-right.svg'
-import React from 'react';
+import React from 'react'
 import { connect } from 'react-redux'
 
 function WorksByFollowedArtists(props) {
