+++ desktop/apps/auction2/routes.js
@@ -3,7 +3,6 @@ import ArticlesQuery from './queries/articles'
 import Articles from '../../collections/articles.coffee'
 import Auction from '../../models/auction.coffee'
 import footerItems from './footer_items'
-import positronql from '../../lib/positronql.coffee'
 import SaleQuery from './queries/sale'
 import MeQuery from './queries/me'
 import { PREDICTION_URL } from '../../config'
